Backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.4.2
opencv-python==4.8.0
numpy==1.26.0
requests==2.31.0
openai==1.30.1
python-dotenv==1.0.0
mangum==0.17.0  # For AWS Lambda integration